ARROW-679: [Format] Change FieldNode, RecordBatch lengths to long, remove LargeRecordBatch. Refactoring
This enables me to delete a bunch of code without losing functionality. C++ users must explicitly opt-in to writing size over INT32_MAX.

cc @julienledem. I have not added checks in Java about sizes over INT32_MAX, wasn't sure where you might want to do that.

Author: Wes McKinney <wes.mckinney@twosigma.com>

Closes #417 from wesm/ARROW-679 and squashes the following commits:

ea237b1 [Wes McKinney] Document allow_64bit for WriteRecordBatch
e237d4a [Wes McKinney] Change FieldNode, RecordBatch lengths to long, remove LargeRecordBatch. Refactoring

+++ java/vector/src/main/java/org/apache/arrow/vector/stream/MessageSerializer.java
@@ -207,7 +207,7 @@ public class MessageSerializer {
     List<ArrowFieldNode> nodes = new ArrayList<>();
     for (int i = 0; i < nodesLength; ++i) {
       FieldNode node = recordBatchFB.nodes(i);
-      nodes.add(new ArrowFieldNode(node.length(), node.nullCount()));
+      nodes.add(new ArrowFieldNode((int)node.length(), (int)node.nullCount()));
     }
     List<ArrowBuf> buffers = new ArrayList<>();
     for (int i = 0; i < recordBatchFB.buffersLength(); ++i) {
@@ -216,7 +216,7 @@ public class MessageSerializer {
       buffers.add(vectorBuffer);
     }
     ArrowRecordBatch arrowRecordBatch =
-        new ArrowRecordBatch(recordBatchFB.length(), nodes, buffers);
+        new ArrowRecordBatch((int)recordBatchFB.length(), nodes, buffers);
     body.release();
     return arrowRecordBatch;
   }
